+++ CHANGELOG.md
@@ -51,6 +51,7 @@ Version 0.7.0-SNAPSHOT - Build 20160108
 - KnowledgeBase is no longer serializable. Its serializable fields are stored individually into the Database.
 - Restructuring the framework to remove all FindBug warnings.
 - KnowledgeBase is now an interface, while the implementation moved to StandardKnowledgeBase. The interface contains a static factory method to produce any KB. This enable us to define the knowledgeBase field of BaseTrainable private and final.
+- Moved SensitivityRates and StepwiseCompatible to the machinelearning/common/ package.
 
 Version 0.6.1 - Build 20160102
 ------------------------------